simplify the package structure, rename AppEntry to Main as we put int controller methods there
