--- src/main/java/demo/todo/jfinal/model/Account.java
@@ -1,12 +1,0 @@
-package demo.todo.jfinal.model;
-
-
-import demo.todo.jfinal.model.base.BaseAccount;
-
-/**
- * Generated by JFinal.
- */
-@SuppressWarnings("serial")
-public class Account extends BaseAccount<Account> {
-	public static final Account dao = new Account().dao();
-}
