--- src/main/java/demo/todo/jfinal/model/base/BaseAccount.java
@@ -1,60 +1,0 @@
-package demo.todo.jfinal.model.base;
-
-import com.jfinal.plugin.activerecord.IBean;
-import com.jfinal.plugin.activerecord.Model;
-
-/**
- * Generated by JFinal, do not modify this file.
- */
-@SuppressWarnings("serial")
-public abstract class BaseAccount<M extends BaseAccount<M>> extends Model<M> implements IBean {
-
-	public void setId(Integer id) {
-		set("id", id);
-	}
-
-	public Integer getId() {
-		return get("id");
-	}
-
-	public void setName(String name) {
-		set("name", name);
-	}
-
-	public String getName() {
-		return get("name");
-	}
-
-	public void setPassword(String password) {
-		set("password", password);
-	}
-
-	public String getPassword() {
-		return get("password");
-	}
-
-	public void setDisable(Integer disable) {
-		set("disable", disable);
-	}
-
-	public Integer getDisable() {
-		return get("disable");
-	}
-
-	public void setCreatedTime(java.util.Date createdTime) {
-		set("createdTime", createdTime);
-	}
-
-	public java.util.Date getCreatedTime() {
-		return get("createdTime");
-	}
-
-	public void setRuleId(Integer ruleId) {
-		set("rule_id", ruleId);
-	}
-
-	public Integer getRuleId() {
-		return get("rule_id");
-	}
-
-}
