+++ src/main/resources/jfinal/demo/Main/list.html
@@ -1,3 +1,3 @@
-#for(account:accounts)
+#for(account:result)
     #(account.name)<br>
 #end
